refactor(cli): extract confirmAction to utils/prompt.ts — eliminate duplication between migrate and fields
