+++ packages/cli/src/commands/fields.ts
@@ -2,6 +2,7 @@ import { Command } from 'commander'
 import { resolveClient } from '../config/resolve-context.js'
 import { formatOutput } from '../output/formatter.js'
 import { isJsonMode } from '../program.js'
+import { confirmAction } from '../utils/prompt.js'
 import type { GlobalFlags } from '../types.js'
 
 const DESTRUCTIVE_FIELDS_DELETE_ERROR = (entity: string, fieldName: string) => ({
@@ -100,20 +101,3 @@ export function registerFieldsCommand(program: Command): void {
     })
 }
 
-async function confirmAction(prompt: string): Promise<boolean> {
-  return new Promise((resolve, reject) => {
-    process.stderr.write(prompt)
-    process.stdin.setEncoding('utf8')
-    const onData = (data: Buffer | string) => { cleanup(); resolve(data.toString().trim().toLowerCase() === 'y') }
-    const onEnd = () => { cleanup(); resolve(false) }
-    const onError = (err: Error) => { cleanup(); reject(err) }
-    const cleanup = () => {
-      process.stdin.removeListener('data', onData)
-      process.stdin.removeListener('end', onEnd)
-      process.stdin.removeListener('error', onError)
-    }
-    process.stdin.once('data', onData)
-    process.stdin.once('end', onEnd)
-    process.stdin.once('error', onError)
-  })
-}
