+++ packages/cli/src/commands/migrate.ts
@@ -2,6 +2,7 @@ import { Command } from 'commander'
 import { resolveClient } from '../config/resolve-context.js'
 import { CliValidationError } from '../errors.js'
 import { isJsonMode } from '../program.js'
+import { confirmAction } from '../utils/prompt.js'
 import type { GlobalFlags } from '../types.js'
 
 const DESTRUCTIVE_ERROR = {
@@ -122,20 +123,3 @@ async function runMigrate(
   }
 }
 
-async function confirmAction(prompt: string): Promise<boolean> {
-  return new Promise((resolve, reject) => {
-    process.stderr.write(prompt)
-    process.stdin.setEncoding('utf8')
-    const onData = (data: Buffer | string) => { cleanup(); resolve(data.toString().trim().toLowerCase() === 'y') }
-    const onEnd = () => { cleanup(); resolve(false) }
-    const onError = (err: Error) => { cleanup(); reject(err) }
-    const cleanup = () => {
-      process.stdin.removeListener('data', onData)
-      process.stdin.removeListener('end', onEnd)
-      process.stdin.removeListener('error', onError)
-    }
-    process.stdin.once('data', onData)
-    process.stdin.once('end', onEnd)
-    process.stdin.once('error', onError)
-  })
-}
